requirements.txt
appdirs==1.4.4
asgiref==3.3.1
autopep8==1.5.5
certifi==2020.12.5
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-ckeditor==6.0.0
django-crispy-forms==1.11.1
django-environ==0.4.5
django-js-asset==1.2.2
django-on-heroku==1.1.2
filelock==3.0.12
gunicorn==20.0.4
psycopg2-binary==2.8.6
pycodestyle==2.6.0
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.4.2
virtualenv-clone==0.5.4
whitenoise==5.2.0
